+++ src/containers/Application.js
@@ -27,6 +27,7 @@ import ReduxService from '../services/ReduxService';
 import ConfigService from '../config/ConfigService';
 import HeaderService from '../services/HeaderService';
 import FooterService from '../services/FooterService';
+import SidebarService from './Sidebar/service';
 
 import TenantSelectionCard from '../modules/tenant/selector/TenantSelectionCard';
 
@@ -91,6 +92,7 @@ class Application extends Component {
 
 		this.HeaderService = new HeaderService(this, "HeaderService");
 		this.FooterService = new FooterService(this, "FooterService");
+		this.SidebarService = new SidebarService(this, "SidebarService");
 
 		this.ReduxService.addReducer("alerts", alertsReducer);
 		this.ReduxService.addReducer("advmode", advancedModeReducer);